NeighborlyAPI/requirements.txt
azure-functions==1.2.1
pymongo==3.10.1
dnspython==2.6.1
